[del] remove wavelet code legacy
