+++ pisap/__init__.py
@@ -15,8 +15,5 @@ that allows sparse decomposition, denoising and deconvolution.
 from .info import __version__
 from pisap.base import io
 from pisap.base.image import Image
-from pisap.base.transform import WaveletTransform
-from pisap.base.wavelet import Wavelet
-from pisap.base.wavelet import WMAP
 import pisap.extensions
