+++ pisap/numerics/noise.py
@@ -118,11 +118,4 @@ def sigma_mad_sparse(grad_op, linear_op):
     sigma: list
         a list of str estimate for each scale.
     """
-    if not isinstance(linear_op, Dictionary): # to preserve code legacy
-        residuals = linear_op.op(grad_op.grad).to_cube()
-        sigma = []
-        for scale_data in residuals[:, 0]:
-            sigma.append(sigma_mad(scale_data))
-        return sigma
-    else:
-        return [sigma_mad(scale_data) for scale_data in linear_op.op(grad_op.grad)]
+    return [sigma_mad(scale_data) for scale_data in linear_op.op(grad_op.grad)]
