--- pisap/sparsity/__init__.py
@@ -1,11 +1,0 @@
-##########################################################################
-# XXX - Copyright (C) XXX, 2017
-# Distributed under the terms of the CeCILL-B license, as published by
-# the CEA-CNRS-INRIA. Refer to the LICENSE file or to
-# http://www.cecill.info/licences/Licence_CeCILL-B_V1-en.html
-# for details.
-##########################################################################
-
-"""
-This module defines Python implementation of the ISAP C++ library.
-"""
